updates to structs
