+++ assets/scripts/update_deps.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+# Get a list of all installed packages
+pip --disable-pip-version-check list --outdated --format=json | python -c "import json, sys; print('\n'.join([x['name'] for x in json.load(sys.stdin)]))" | xargs -n1 pip install -U
+
