+++ worker/jobs.py
@@ -12,11 +12,10 @@ from pymongo.mongo_client import MongoClient
 import numpy as np
 import pandas as pd
 
-from worker.shared.shared import BaseClass
-from worker.shared.database import MongoDbConnector
-from worker.data_model import UtcComparison, SimulationError, UtcSpeciesComparison
-from worker.io import get_sbml_species_names, get_sbml_model_file_from_archive, read_report_outputs
-from worker.output_data import generate_biosimulator_utc_outputs, _get_output_stack
+from shared import BaseClass, MongoDbConnector
+from data_model import UtcComparison, SimulationError, UtcSpeciesComparison
+from io import get_sbml_species_names, get_sbml_model_file_from_archive, read_report_outputs
+from output_data import generate_biosimulator_utc_outputs, _get_output_stack
 
 
 DB_TYPE = "mongo"  # ie: postgres, etc
